Changed size type

+++ src/include/BitStream.h
@@ -22,8 +22,8 @@ private:
     bufflen_t mHead;
     byte* mBuffer;
 
-    void WriteBits(byte data, size_t size);
-    void WriteBits(const void *data, size_t size);
+    void WriteBits(byte data, uint32_t size);
+    void WriteBits(const void *data, uint32_t size);
 
 public:
     explicit OutputBitStream(bufflen_t capacity): mBuffer(nullptr), mHead(0), mCapacity(0)
@@ -43,7 +43,7 @@ public:
     bufflen_t GetSize() {return mHead;}
 
     template <class T>
-    void Write(T data, size_t size = sizeof(T) * BITS_PER_BYTE)
+    void Write(T data, uint32_t size = sizeof(T) * BITS_PER_BYTE)
     {
         static_assert(std::is_arithmetic<T>::value || std::is_enum<T>::value,
                       "Generic BitStream::Write supports only arithmetic or enum types");
@@ -60,8 +60,8 @@ private:
     byte* mBuffer;
 
 
-    void ReadBits(byte &data, size_t size);
-    void ReadBits(void* data, size_t size);
+    void ReadBits(byte &data, uint32_t size);
+    void ReadBits(void* data, uint32_t size);
 
 public:
     explicit InputBitStream(const byte* buffer, bufflen_t size): mHead(0)
@@ -73,7 +73,7 @@ public:
     ~InputBitStream(){ std::free(mBuffer); }
 
     template <class T>
-    void Read(T& data, size_t size = sizeof(T) * BITS_PER_BYTE)
+    void Read(T& data, uint32_t size = sizeof(T) * BITS_PER_BYTE)
     {
         static_assert(std::is_arithmetic<T>::value || std::is_enum<T>::value,
                       "Generic BitStream::Read supports only arithmetic or enum types");
